backend/fastapi/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
python-multipart==0.0.6
Pillow==10.1.0
pycryptodome
